requirements.txt
alabaster==0.7.13
appnope==0.1.3
attrs==22.2.0
Babel==2.12.1
backcall==0.2.0
bleach==6.0.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.2.1
dask==2022.2.0
debugpy==1.6.6
decorator==5.1.1
docutils==0.19
entrypoints==0.4
et-xmlfile==1.1.0
Fiona==1.9.2
fsspec==2023.1.0
geopandas==0.10.2
idna==3.4
imagesize==1.4.1
importlib-metadata==6.1.0
ipykernel==6.16.2
ipython==7.34.0
ipywidgets==8.0.6
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter-client==7.4.9
jupyter-core==4.12.0
jupyterlab-widgets==3.0.7
locket==1.0.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
munch==2.5.0
nest-asyncio==1.5.6
numpy==1.21.6
openpyxl==3.1.2
packaging==23.0
pandas==1.1.5
parso==0.8.3
partd==1.3.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pygeos==0.14
Pygments==2.14.0
pyproj==3.2.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
requests==2.28.2
Rtree==1.0.1
scikit-learn==1.0.2
scipy==1.7.3
shapely==2.0.1
six==1.16.0
sklearn==0.0.post2
snowballstemmer==2.2.0
sphinx==5.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
swifter==1.3.4
threadpoolctl==3.1.0
toolz==0.12.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing-extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
webencodings==0.5.1
widgetsnbextension==4.0.7
xlrd==2.0.1
zipp==3.15.0
